# openapi.yaml
openapi: 3.0.3
info:
  title: TCSS 460 Group 5 API
  description: |
    **Team:** Group 5
    **Deployed URL:** https://tcss460-group-5-api.onrender.com
    A RESTful API for the TCSS 460 group project.
  version: 0.1.0
servers:
  - url: https://tcss460-group-5-api.onrender.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/tv/{id}:
    get:
      tags: [TV]
      summary: Get a TV show by ID
      description: Returns detailed information for a specific TV show by its TMDB ID.
      parameters:
        - name: id
          in: path
          description: The TMDB TV show ID.
          required: true
          schema:
            type: integer
            minimum: 1
          example: 1396
      responses:
        '200':
          description: Detailed information about the requested TV show.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TvDetail'
        '400':
          description: Invalid TV show ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: TV show id must be a positive integer
        '404':
          description: TV show not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: TV show with id 99999 not found
        '500':
          description: Internal server error or TMDB API failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TvDetail:
      type: object
      description: Detailed TV show information returned from the single TV endpoint.
      properties:
        name:
          type: string
          example: Breaking Bad
        original_name:
          type: string
          example: Breaking Bad
        overview:
          type: string
          example: A chemistry teacher diagnosed with cancer turns to manufacturing drugs.
        tagline:
          type: string
          nullable: true
          example: Change the equation.
        first_air_date:
          type: string
          format: date
          example: '2008-01-20'
        last_air_date:
          type: string
          format: date
          nullable: true
          example: '2013-09-29'
        status:
          type: string
          example: Ended
        type:
          type: string
          example: Scripted
        adult:
          type: boolean
          example: false
        genres:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
          example: [{ id: 18, name: Drama }]
        original_language:
          type: string
          example: en
        spoken_languages:
          type: array
          items:
            type: object
            properties:
              english_name:
                type: string
              iso_639_1:
                type: string
              name:
                type: string
        origin_country:
          type: array
          items:
            type: string
          example: ['US']
        number_of_episodes:
          type: integer
          example: 62
        number_of_seasons:
          type: integer
          example: 5
        seasons:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              episode_count:
                type: integer
              air_date:
                type: string
                format: date
                nullable: true
        episode_run_time:
          type: array
          items:
            type: integer
          example: [45]
        last_episode_to_air:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            name:
              type: string
            episode_number:
              type: integer
            season_number:
              type: integer
        next_episode_to_air:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            name:
              type: string
            episode_number:
              type: integer
            season_number:
              type: integer
        networks:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              logo_path:
                type: string
                nullable: true
              origin_country:
                type: string
        production_companies:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              logo_path:
                type: string
                nullable: true
              origin_country:
                type: string
        production_countries:
          type: array
          items:
            type: object
            properties:
              iso_3166_1:
                type: string
              name:
                type: string
        created_by:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              profile_path:
                type: string
                nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: A human-readable error message.
          example: Not found
...